Introduce Loga::Event to avoid 💣
Currently when the hash is passed as a message to the formatter it must respect a contract.
```ruby
logger.info(
  message:       'Hello World',  # REQUIRED
  type:          'cron',         # OPTIONAL
  timestamp:     Time,           # OPTIONAL
  event:         {               # OPTIONAL
    'color' =>   'red',
  },
  exception:     Exception,      # OPTIONAL
)
```
Should the hash not honour the contract, the formatter will go :boom:.

This is risky in legacy applications as they may log arbitrary hash structures.

To eliminate this risk we introduce `Loga::Event`.
```ruby
event = Loga::Event.new(...)
logger.info(event)
```

With that the formatter can handle arbitrary inputs responding to `to_s` and instances of `Loga:Event`