:see_no_evil: Improve Sinatra/Rack request logs when format is set to simple
This is relevant in Sinatra application when the Loga format is set to simple.
Then the Loga::Rack::Logger passes an instance of Loga::Event to Loga.logger which
uses the ActiveSupport::Logger::SimpleFormatter.

ActiveSupport::Logger::SimpleFormatter calls inspect when receiving an non-String
object. Overriding inspect in Loga::Event improves the output.