Support rails 6
This commit modifies the railtie for loga to support rails 6. Our monkey patch for Rails::Rack::Logger is still valid as there has been no changes for this file between rails 5.2 and rails 6.0 in the upstream.

This commit also ensures the test suite runs against rails 6 for ruby versions greater than or equal to 2.5 which is the ruby version required for rails 6.

This should help users who are upgrading to rails 6 will be coming across a "Loga::ConfigurationError: Rails 6 is unsupported" error.