Rework middleware to be stateless, to avoid race conditions
A middleware is initialized once, therefore reused across multiple
requests. This is a problem when using Puma, since it uses threads. Any
state within the middleware is in danger of being overwritten by another
request. I've killed off all the state in the middleware to prevent
these issues

I've also reworked the spec a little bit to remove all the stubbing,
including of the actual test subject