[USM-1820] Add Redis to CI for integration testing Sidekiq 7
Sidekiq's config API changed in v7 so you no longer pass in a Redis object, and it uses redis-client rather than redis-rb to interface with Redis, so the approach for intergration tests of pre-7 versions using Fakeredis doesn't work.  It'd be possible to stub out the connection at a lower level, but after discussion we think it's more useful to test against the real Redis.